backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
boto3==1.34.25
PyMuPDF==1.23.18
pdfplumber==0.10.4
pytesseract==0.3.10
Pillow==10.2.0
scikit-learn==1.4.0
pandas==2.2.0
python-dotenv==1.0.1
httpx==0.26.0
openai==1.58.1
jinja2==3.1.3
weasyprint==61.0
reportlab==4.1.0
